requirements.txt
mcp>=1.0.0
caldav>=1.3.0
icalendar>=5.0.0
pydantic>=2.4.0
httpx>=0.27.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0